% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{india}
\alias{india}
\title{Replication Data for: Causal Inference with Interference and Noncompliance in Two-Stage Randomized Experiments.}
\format{
A data frame with columns:
\describe{
 \item{id}{The id for the village.}
 \item{DistrictId}{The id for the district.}
 \item{Z}{The treatment status for the individual.}
 \item{A}{The treatment assignment mechanism.}
 \item{D}{Whether or not the individual enrolled.}
 \item{Y}{The hospital expenditure.}
 \item{X}{Enumeration of the patients.}
}
}
\source{
\url{https://doi.org/10.7910/DVN/N7D9LS}
}
\usage{
data(india)
}
\description{
Replication Data for: Causal Inference with Interference and Noncompliance in Two-Stage Randomized Experiments.
}
\keyword{datasets}
