% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_Methods.R
\name{print.regression}
\alias{print.regression}
\title{Print Method for the RCT2 Package}
\usage{
\method{print}{regression}(x, ...)
}
\arguments{
\item{x}{A list object generated by running one of the analyses on a data set.}

\item{...}{ignored}
}
\value{
\code{NULL}
}
\description{
This function prints a nicely formatted summary of the three functions in the RCT2 package.
}
\details{
For the details of the method implemented by this function, see the
references.
}
\references{
Kosuke Imai, Zhichao Jiang and Anup Malani (2018).
\dQuote{Causal Inference with Interference and Noncompliance in the Two-Stage Randomized Experiments}, \emph{Technical Report}. Department of Politics, Princeton
University.
}
\author{
Kosuke Imai, Department of Statistics, Harvard University
\email{imai@harvard.edu}, \url{https://imai.fas.harvard.edu/};
Zhichao Jiang, School of Public Health and Health Sciences, University of Massachusetts Amherst
\email{zhichaojiang@umass.edu};
Karissa Huang, Department of Statistics, Harvard College
\email{krhuang@college.harvard.edu}
}
\keyword{experiments}
\keyword{randomized}
\keyword{two-stage}
