% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YSRE.R
\name{YSRE}
\alias{YSRE}
\title{Sample size calculations for detecting a specific alternative}
\usage{
YSRE(J, var)
}
\arguments{
\item{J}{A vector of cluster labels for each observation of the dataset.}

\item{var}{A vector of the variable of interest, from which the binary treatment vector and outcome vector are to be extracted.}
}
\value{
A list which contains the following item:
\item{Y}{ A list of the outcome variable for each observation. }
}
\description{
This function extracts the necessary outcome vector `Y`  for further analysis.
For the details of the method implemented by this function, see the
references.
}
\references{
Zhichao Jiang, Kosuke Imai (2020).
\dQuote{Statistical Inference and Power Analysis for Direct and Spillover Effects in Two-Stage Randomized Experiments}, \emph{Technical Report}.
}
\author{
Kosuke Imai, Department of Statistics, Harvard University
\email{imai@harvard.edu}, \url{https://imai.fas.harvard.edu/};
Zhichao Jiang, School of Public Health and Health Sciences, University of Massachusetts Amherst
\email{zhichaojiang@umass.edu};
Karissa Huang, Department of Statistics, Harvard College
\email{krhuang@college.harvard.edu}
}
\keyword{experiments}
\keyword{randomized}
\keyword{two-stage}
