% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZSRE.R
\name{ZSRE}
\alias{ZSRE}
\title{Sample size calculations for detecting a specific alternative}
\usage{
ZSRE(assign, J)
}
\arguments{
\item{assign}{The assignment vector for each observation of the data.}

\item{J}{A vector of cluster labels for each observation of the dataset.}
}
\value{
A list which contains the following item:
\item{Z}{ A  list of the binary treatment assignment variable for each observation. }
}
\description{
This function extracts the binary treatment assignment vector `Z`.
For the details of the method implemented by this function, see the
references.
}
\references{
Zhichao Jiang, Kosuke Imai (2020).
\dQuote{Statistical Inference and Power Analysis for Direct and Spillover Effects in Two-Stage Randomized Experiments}, \emph{Technical Report}.
}
\author{
Kosuke Imai, Department of Statistics, Harvard University
\email{imai@harvard.edu}, \url{https://imai.fas.harvard.edu/};
Zhichao Jiang, School of Public Health and Health Sciences, University of Massachusetts Amherst
\email{zhichaojiang@umass.edu};
Karissa Huang, Department of Statistics, Harvard College
\email{krhuang@college.harvard.edu}
}
\keyword{experiments}
\keyword{randomized}
\keyword{two-stage}
