% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_Methods.R
\name{print}
\alias{print}
\alias{print.random}
\alias{print.regression}
\alias{print.parametric}
\title{Print Method for the RCT2 Package}
\usage{
print(x)

\method{print}{random}(x)

\method{print}{regression}(x)

\method{print}{parametric}(x)
}
\arguments{
\item{x}{A list object generated by running one of the analyses on a data set.}
}
\value{
\code{NULL}
}
\description{
This function prints a nicely formatted summary of the three functions in the RCT2 package.
}
\details{
For the details of the method implemented by this function, see the
references.
}
\references{
Kosuke Imai, Zhichao Jiang and Anup Malani (2018).
\dQuote{Causal Inference with Interference and Noncompliance in the Two-Stage Randomized Experiments}, \emph{Technical Report}. Department of Politics, Princeton
University.
}
\author{
Kosuke Imai, Department of Politics, Princeton University
\email{kimai@Princeton.Edu}, \url{http://imai.princeton.edu};
Zhichao Jiang, Department of Politics, Princeton University
\email{zhichaoj@princeton.edu}.
}
\keyword{experiments}
\keyword{randomized}
\keyword{two-stage}
